task 4

+++ vttp2023-batch4-ssf-assessment/src/main/java/vttp/ssf/assessment/eventmanagement/models/Event.java
@@ -9,13 +9,13 @@ public class Event {
     private Integer eventId;
     private String eventName;
     private Integer eventSize;
-    private Integer eventDate;
+    private Long eventDate;
     private Integer participants;
 
     public Event() {
     }
 
-    public Event(Integer eventId, String eventName, Integer eventSize, Integer eventDate, Integer participants) {
+    public Event(Integer eventId, String eventName, Integer eventSize, Long eventDate, Integer participants) {
         this.eventId = eventId;
         this.eventName = eventName;
         this.eventSize = eventSize;
@@ -61,7 +61,7 @@ public class Event {
         return dt;
     }
 
-    public void setEventDate(Integer eventDate) {
+    public void setEventDate(Long eventDate) {
         this.eventDate = eventDate;
     }
